[FEC] Make RS codec generally available
